+++ protocols/request-response/tests/error_reporting.rs
@@ -50,7 +50,7 @@ async fn report_outbound_failure_on_read_response() {
             .send_request(&peer1_id, Action::FailOnReadResponse);
 
         let (peer, req_id_done, error) = wait_outbound_failure(&mut swarm2).await.unwrap();
-        assert_eq!(peer, peer1_id);
+        assert_eq!(peer, Some(peer1_id));
         assert_eq!(req_id_done, req_id);
 
         let error = match error {
@@ -93,7 +93,7 @@ async fn report_outbound_failure_on_write_request() {
             .send_request(&peer1_id, Action::FailOnWriteRequest);
 
         let (peer, req_id_done, error) = wait_outbound_failure(&mut swarm2).await.unwrap();
-        assert_eq!(peer, peer1_id);
+        assert_eq!(peer, Some(peer1_id));
         assert_eq!(req_id_done, req_id);
 
         let error = match error {
@@ -150,7 +150,7 @@ async fn report_outbound_timeout_on_read_response() {
             .send_request(&peer1_id, Action::TimeoutOnReadResponse);
 
         let (peer, req_id_done, error) = wait_outbound_failure(&mut swarm2).await.unwrap();
-        assert_eq!(peer, peer1_id);
+        assert_eq!(peer, Some(peer1_id));
         assert_eq!(req_id_done, req_id);
         assert!(matches!(error, OutboundFailure::Timeout));
     };
@@ -202,7 +202,7 @@ async fn report_outbound_failure_on_max_streams() {
             .send_request(&peer1_id, Action::FailOnMaxStreams);
 
         let (peer, req_id_done, error) = wait_outbound_failure(&mut swarm2).await.unwrap();
-        assert_eq!(peer, peer1_id);
+        assert_eq!(peer, Some(peer1_id));
         assert_eq!(req_id_done, outbound_req_id);
         assert!(matches!(error, OutboundFailure::Io(_)));
     };
@@ -235,7 +235,7 @@ async fn report_inbound_failure_on_read_request() {
             .send_request(&peer1_id, Action::FailOnReadRequest);
 
         let (peer, req_id_done, error) = wait_outbound_failure(&mut swarm2).await.unwrap();
-        assert_eq!(peer, peer1_id);
+        assert_eq!(peer, Some(peer1_id));
         assert_eq!(req_id_done, req_id);
 
         match error {
@@ -294,7 +294,7 @@ async fn report_inbound_failure_on_write_response() {
             .send_request(&peer1_id, Action::FailOnWriteResponse);
 
         let (peer, req_id_done, error) = wait_outbound_failure(&mut swarm2).await.unwrap();
-        assert_eq!(peer, peer1_id);
+        assert_eq!(peer, Some(peer1_id));
         assert_eq!(req_id_done, req_id);
 
         match error {
@@ -351,7 +351,7 @@ async fn report_inbound_timeout_on_write_response() {
             .send_request(&peer1_id, Action::TimeoutOnWriteResponse);
 
         let (peer, req_id_done, error) = wait_outbound_failure(&mut swarm2).await.unwrap();
-        assert_eq!(peer, peer1_id);
+        assert_eq!(peer, Some(peer1_id));
         assert_eq!(req_id_done, req_id);
 
         match error {
@@ -611,7 +611,7 @@ async fn wait_inbound_failure(
 
 async fn wait_outbound_failure(
     swarm: &mut Swarm<request_response::Behaviour<TestCodec>>,
-) -> Result<(PeerId, OutboundRequestId, OutboundFailure)> {
+) -> Result<(Option<PeerId>, OutboundRequestId, OutboundFailure)> {
     loop {
         match swarm.select_next_some().await.try_into_behaviour_event() {
             Ok(request_response::Event::OutboundFailure {